+++ db/migrations/inventory/20260421162704_add_reserved_column.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 -- +goose StatementBegin
 ALTER TABLE products
-ADD COLUMN reserved INTEGER NOT NULL DEFAULT 0,
+ADD COLUMN reserved BIGINT NOT NULL DEFAULT 0,
 ADD CONSTRAINT check_reserved_not_exceed_quantity CHECK (reserved <= quantity),
 ADD CONSTRAINT check_reserved_positive CHECK (reserved >= 0);
